test: move issue tests to subfolder
