+++ packages/core/test/issues/issue-261.test.js
@@ -1,13 +1,13 @@
 "use strict";
 
-var Processor = require("../processor.js");
+var Processor = require("../../processor.js");
 
 describe("/issues", function() {
     describe("/261", function() {
         it("should allow colons in rules that also use :external()", function() {
             var processor = new Processor();
             
-            return processor.file("./packages/core/test/specimens/issues/261/2.css")
+            return processor.file(require.resolve("./specimens/261/2.css"))
             .then(() => processor.output())
             .then((result) => expect(result.css).toMatchSnapshot());
         });
