+++ packages/core/test/issues/issue-56.test.js
@@ -2,26 +2,24 @@
 
 var dedent = require("dedent"),
     namer  = require("test-utils/namer.js"),
-
-    Processor = require("../processor.js");
+    
+    Processor = require("../../processor.js");
 
 describe("/issues", function() {
-    describe("/24", function() {
-        it("should be able to compose using a value", function() {
+    describe("/56", function() {
+        it("should prune rules that only compose, but leave them in the exports", function() {
             var processor = new Processor({
                     namer
                 });
             
             return processor.string(
-                "./packages/core/test/specimens/composition.css",
-                dedent(`
-                    @value simple: "./simple.css";
-                    
-                    .a {
-                        composes: wooga from simple;
-                        background: #000;
-                    }
-                `)
+                    "./packages/core/test/specimens/issues/56.css",
+                    dedent(`
+                        .booga { color: red }
+                        .fooga { composes: booga }
+                        .fooga:hover { color: blue }
+                        .wooga { composes: booga }
+                    `)
             )
             .then((result) => {
                 expect(result.exports).toMatchSnapshot();
