--- packages/core/test/specimens/issues/105/1.css
@@ -1,1 +1,0 @@
-.wooga { composes: fooga from "./symlink.css"; }
