--- packages/core/test/specimens/issues/105/folder/2.css
@@ -1,1 +1,0 @@
-.fooga { color: red; }
