--- packages/core/test/specimens/issues/58/1.css
@@ -1,3 +1,0 @@
-.other1 { color: red; }
-.other2 { color: navy; }
-.other3 { color: blue; }
